Moving FeatureBinding and CompositeBinding into android.internal package.
