+++ formula-android/src/main/java/com/instacart/formula/integration/Binding.kt
@@ -4,7 +4,7 @@ import com.instacart.formula.FormulaContext
 import com.instacart.formula.android.FlowFactory
 import com.instacart.formula.android.internal.ComponentFactory
 import com.instacart.formula.fragment.FragmentEnvironment
-import com.instacart.formula.integration.internal.CompositeBinding
+import com.instacart.formula.android.internal.CompositeBinding
 
 /**
  * Defines how specific keys bind to the state management associated
